fix: text on map in dark mode
